requirements.txt
langchain
langchain-core
langchain-openai
openai
python-dotenv
pypdf
langchain-community
langchain-tavily
tavily